move spectrum to mct_state
